deleted all files, components and modules that are not related to Authentication
a note on onboarding 📝:
there is a problem which is that a member jwt will be marked with
onboarding complete, before the user actually gets added to the DB.
the best way to do it is
1. call clerk and update the metadata (without setting onboarding)
2. ask clerk to reload.
3. create the member in DB
4. call clerk and upddate the metadata setting onboarding to true
5. ask clerk to reload (so that the frontend also know that onboarding
was complete)
---
However doing this requries 3 additional calls to clerk which
significantly increases delay.

my initial idea was to asyncrounsly set onboarding, but then the
frontend would be stuck on the on-boarding form.